+++ secureli/actions/action.py
@@ -12,7 +12,7 @@ from secureli.modules.language_analyzer import language_analyzer, language_suppo
 from secureli.modules.core.core_services.scanner import ScannerService
 from secureli.modules.core.core_services.updater import UpdaterService
 
-from secureli.modules.shared.utilities.formatter import format_sentence_list
+from secureli.modules.shared.utilities import format_sentence_list
 
 
 class ActionDependencies:
