+++ secureli/modules/shared/models/echo.py
@@ -15,3 +15,17 @@ class Color(str, Enum):
     MAGENTA = "magenta"
     CYAN = "cyan"
     WHITE = "white"
+
+
+class Level(str, Enum):
+    debug = "DEBUG"
+    info = "INFO"
+    warn = "WARN"
+    error = "ERROR"
+    off = "OFF"
+
+    def __str__(self) -> str:
+        return self.value
+
+    def __repr__(self) -> str:
+        return self.__str__()
