+++ tests/modules/shared/utilities/test_logging.py
@@ -1,13 +1,13 @@
-from secureli.modules.shared.utilities.logging import EchoLevel
+from secureli.modules.shared.models.echo import Level
 
 
 def test_that_echo_level_str_returns_enum_val():
-    level = EchoLevel.info
+    level = Level.info
 
     assert str(level) == level.value
 
 
 def test_that_echo_level_repr_returns_str_implementation():
-    level = EchoLevel.debug
+    level = Level.debug
 
     assert repr(level) == str(level)
